Container startup improvements

+++ devstack/resources/photonic/photonic.sh
@@ -1,32 +1,37 @@
 #!/bin/bash
 
-cd /opt/tachyonic/luxon
-python3 setup.py develop
+INSTALLED=/installed
 
-cd /opt/tachyonic/psychokinetic
-python3 setup.py develop
+if [ ! -f $INSTALLED ]; then
+    cd /opt/tachyonic/luxon
+    python3 setup.py develop
 
-cd /opt/tachyonic/infinitystone
-python3 setup.py develop
+    cd /opt/tachyonic/psychokinetic
+    python3 setup.py develop
 
-cd /opt/tachyonic/photonic
-python3 setup.py develop
+    cd /opt/tachyonic/infinitystone
+    python3 setup.py develop
 
-# External Modules...
-cd /opt/tachyonic/tradius
-python3 setup.py develop
+    cd /opt/tachyonic/photonic
+    python3 setup.py develop
 
-# Continued...
+    # External Modules...
+    cd /opt/tachyonic/tradius
+    python3 setup.py develop
 
-mkdir /opt/tachyonic/www/photonic
-cd /opt/tachyonic/www/photonic
-luxon -i photonic .
-luxon -i infinitystone.ui .
-luxon -i tradius.ui .
-ln -s /opt/tachyonic/www/infinitystone/public.pem .
+    # Continued...
 
-rm /etc/nginx/sites-enabled/default
-ln -s /opt/tachyonic/photonic.nginx /etc/nginx/sites-enabled/photonic
+    mkdir /opt/tachyonic/www/photonic
+    cd /opt/tachyonic/www/photonic
+    luxon -i photonic .
+    luxon -i infinitystone.ui .
+    luxon -i tradius.ui .
+    ln -s /opt/tachyonic/www/infinitystone/public.pem .
+
+    rm /etc/nginx/sites-enabled/default
+    ln -s /opt/tachyonic/photonic.nginx /etc/nginx/sites-enabled/photonic
+    touch $INSTALLED
+fi
 
 /etc/init.d/nginx start
